// src/lib/schemas.ts
export const SpotifyAlbumSchema = z.object({
  album_type: z.string(),
  href: z.string().url(),
  id: z.string(),
  name: z.string(),
  release_date: z.string(),
  total_tracks: z.number(),
  type: z.string(),
  images: z.array(z.object({
    height: z.number(),
    url: z.string().url(),
    width: z.number(),
  })),
})